# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api/v1",
    "paths": {
        "/subscriptions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get all subscriptions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user UUID",
                        "name": "user-uuid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "service name",
                        "name": "service-name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.SubscriptionResp"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPErrorResp"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create subscription",
                "parameters": [
                    {
                        "description": "subscription",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateSubscriptionReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "$ref": "#/definitions/main.SubscriptionResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPErrorResp"
                        }
                    }
                }
            }
        },
        "/subscriptions/stats": {
            "get": {
                "description": "Get stats for period (which is a closed interval), user and service filters",
                "produces": [
                    "application/json"
                ],
                "summary": "Get subscriptions stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user UUID",
                        "name": "user-uuid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "service name",
                        "name": "service-name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "period start",
                        "name": "period-start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "period end",
                        "name": "period-end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/main.SubscriptionsStatsResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPErrorResp"
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get subscription by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "subscription id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/main.SubscriptionResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPErrorResp"
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Cancel subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "subscription id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok"
                    },
                    "404": {
                        "description": "No subscription that is not ended yet with such ID",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CreateSubscriptionReq": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "04-2026"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string",
                    "example": "02-2026"
                },
                "user_uuid": {
                    "type": "string",
                    "example": "cd2a0341-669e-48ca-9210-839b943e75ae"
                }
            }
        },
        "main.HTTPErrorResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "main.SubscriptionResp": {
            "type": "object",
            "properties": {
                "cancelation_date": {
                    "type": "string",
                    "example": "03-2026"
                },
                "end_date": {
                    "type": "string",
                    "example": "04-2026"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string",
                    "example": "02-2026"
                },
                "user_uuid": {
                    "type": "string",
                    "example": "cd2a0341-669e-48ca-9210-839b943e75ae"
                }
            }
        },
        "main.SubscriptionsStatsResp": {
            "type": "object",
            "properties": {
                "revenue": {
                    "type": "integer"
                }
            }
        }
    }
}